Introduce IProtocolLayer.
Instead of three separate interfaces, there is one standard interface that
is implemented by all layers in the protocol stack.

This should allow compression and SSL to be treated more uniformly in the
protocol stack.

IConversation is removed as a separate interface, and moved to being a
method on IProtocolLayer.