Explicitly wait for asynchronous method.
The method completes synchronously, so the waiting isn't strictly necessary. However, the new code will rethrow any exception that the returned Task may have captured, which is useful for correctness (not that Dispose should throw an exception). Additionally, a call to `ConfigureAwait` without an `await` is confusing; this new code now uses the "standard" pattern for synchronously calling an async method.