Implement CancellationToken support in Async methods.
Instead of cancelling the current socket operation as soon as possible (which might leave the connection in an unknown state), cancelling the CancellationToken triggers MySqlCommand.Cancel, which executes "KILL QUERY n" against the database, logically terminating the current statement.

This should result in the eventual cancellation of the query and OperationCanceledException being raised on a subsequent call to ReadAsync (or similar). We assume that the same CancellationToken keeps being passed to all DB methods involved in reading a single result set, and that the slight delay in actual cancellation is not significant to the caller.