Dispose objects properly. Fixes #275
This commit doesn't ensure that an ObjectDisposedException will be thrown
if an object is used after it's disposed; a NullReferenceException or
some other exception may be thrown instead. These are internal types
that should always be used correctly in this code, so tracking incorrect
use of Dispose is a low priority.