Limit command cancellation attempts.
In a load-balanced environment, we may not connect to the same server when attempting to cancel an existing command, which means that cancelling will fail. Similarly, in a proxied environment, we may not have the right thread ID to cancel. Since cancellation (which opens another connection to the server) executes synchronously when the CancellationToken is checked, enforce a limit on the number of times it happens so that the canceled reader is not perpetually trying to cancel and failing.