Handle failure to find procedure. Fixes #282
If no parameters are found, assume the schema name is incorrect and don't populate the cache. This will require the user to supply the stored procedure's parameters in the correct order (but that will be successful).

Additionally, skip parameter lookup if the schema name is unknown (since that will always fail).