Prevent the same reader being closed twice. Fixes #423
Dispose(bool) calls DoClose then (on certain frameworks) base.Dispose(bool) calls Close, which calls DoClose again.

This normally wouldn't be a problem except that when a stored procedure has OUT parameters, they are read from a nested MySqlDataReader. This reentrancy allows a reader to be closed twice if an exception is thrown, which masks the underlying exception and throws InvalidOperationExcepion "Expected state to be Failed but was Connected" instead.