Add a client parameter to specify SSL version(s).
Prior to this patch, TLS1.2 was always disabled on Windows client, and
enabled elsewhere,  but this was not correct.

The ability to use TLS1.2 depends entirely on the way how *server* was
built.

If server was built with YASSL, TLS1.2 negotation would fail due to YaSSL
bug described here https://jira.mariadb.org/browse/MDEV-12190
(no graceful version downgrade)

If the server was built with OpenSSL, TLS1.2 negotiation would succeed,
or gracefully downgrades for older versions of OpenSSL.

FIX:
This patch allows users to specify the SSL version(s) in the connection
string with SslProtocols parameter. Default value for it,
which works with all versions of server, is TLS|TLS11.
User can overwrite it with e.g TLS12, if required.