auth_gssapi - add optional ServerSPN parameter for service principle name.
This parameter is necessary to support mutual authentication feature of the
Kerberos protocol, i.e client can verify server's identity with it.

The parameter can not be used in NTLM scenarios. If Negotiate protocol
falls back to NTLM, and ServerSPN is set, AuthenticationException will be
thrown, because NTLM can not be used to verify server's identity.

This ServerSPN can also be used in some cases to connect to server, if
server does not report correct SPN to clients.


Signed-off-by: Vladislav Vaintroub <wlad@mariadb.com>