Initial implementation of batch commands.
Introduced IMySqlCommand abstraction for commands and batch commands. Replaced PreparedStatementCommandExecutor with ICommandPayloadCreator and a new mechanism for executing commands in a loop. MySqlDataReader can iterate over multiple commands and return multiple result sets.

Many features for batched commands are not yet implemented but the basic framework is now in place and can be extended.

Signed-off-by: Bradley Grainger <bgrainger@gmail.com>