Add unit of measure and default to ConnectionTimeout
The summary is what is shown by consumers  when using autocomplete.  This change allows consumers to easily see the unit of measure (seconds) as well as the default value when using autocomplete/code completion.

I'm proposing this change because it wasn't clear to me if the timeout was in seconds or milliseconds. Also, i had no idea what the default is/was until I added a breakpoint in my code that was consuming this.

Signed-off-by: Amalfitano, Joseph <Joseph.Amalfitano@ssa.gov>

~
~