Always set the CLIENT_LOCAL_FILES flag. Fixes #853
If the AllowLoadLocalInfile=True connection string option is specified, then always pass the CLIENT_LOCAL_FILES flag to the server, regardless of whether it was set in the initial handshake packet.

This works around a problem where the Azure Database for MySQL/MariaDB proxy may not specify that flag, but the backend server supports it.