Add [Category] and [DisplayName] to all connection options. Fixes #1030
This aligns the PropertyDescriptor.DisplayName with the DbConnectionStringBuilder key for getting/setting the property.

Fixes ICustomTypeDescriptor-related issues with MySqlConnectionStringBuilder.

Rearranged connection string option documentation to put the DisplayName first in the list of synonyms.

This does introduce breaking changes for anyone who was depending on the exact format of option keys in the rendered connection string. (All previous options are still accepted.)