Reword "recovering leaked sessions" log message.
The previous message implied the possible existence of leaked sessions, and could be confusing/alarming when reading log files. The new message is intended to imply that there could be some leaked sessions (but there may very well be none) and that detecting them will take a small amount of work.