Work around server query attributes bug. Fixes #1106
MySQL Server versions 8.0.23-8.0.25 contain bugs that prevent them from parsing attributes in the COM_STMT_EXECUTE packet. The new PARAMETER_COUNT_AVAILABLE flag is added in 8.0.26 to enable attributes to be sent. If we detect an older server (that advertises support for query attributes), we send an empty string parameter name, but omit the attributes.