Update to PR #1134 based on feedback
Change the logic of the ssl retry so trigger if the user has set
nothing, or the user has set Tls12 and something else.

Also change the fallback settings so that the options selected
must be allowed by the original selection. For example do
not set the fallback to 1.0 or 1.1 if the user originally selected
1.1 and 1.2.

Signed-off-by: Andrew Nagel andrew.nagel@gmail.com
Signed-off-by: Andrew Nagel <andrew.nagel.ctr@dot.gov>