Remove AddContinuation methods.
Instead of splitting the implementation across eight methods, including two static local methods (to try to create an optimal implementation), rely on the compiler creating an optimized async state machine. This reduces the number of async methods from two to one, reducing the number of async state machines that are allocated.